rpi: override create-devena-initrd
There should be TWO devena-firstboot initramfs images for the Raspberry
Pi devices - one using the 4k kernel (for Raspberry Pi 3 and 4), and one
using the 16k kernel (for Raspebrry Pi 5).

Also ship the configuration in order to make the bootloader load the
corresponding initramfs images. The configuration and added include
directive will be removed during the setup.