create-devena-initrd: disable host-only mode
Host-only mode uses host system configuration to generate the initrd
image for the specific host. Despite the hardware the image is going to
be flashed for will be the same, we are still going to disable this.

Apart from drivers specific to the host system, several system
configurations will be copied to this initrd, e.g. /etc/passwd. This
causes the sulogin(8) no longer asking for Enter, instead it asks for
the root password which does not exist on the host system.